Fix hook return value.
